// libs/app/src/lib/states/enums/food-details-status.enum.ts
export enum FoodDetailsStatusEnum {
  pending = 'pending',
  loading = 'loading',
  createSuccess = 'createSuccess',
  loadSuccess = 'loadSuccess',
  updateSuccess = 'updateSuccess',
  deleteSuccess = 'deleteSuccess',
  createError = 'createError',
  deleteError = 'deleteError',
  updateError = 'updateError',
  loadError = 'loadError',
}